Refactor the backing store concept to be a  bit more pluggable
This paves the way for a backing store that could realistically support
a sparse Merkle Tree that is 2^256 leaves wide.

That is what would be needed to use the sparse Merkle Tree code for
something like the use case described in Revocation Transparency or any
other use case where we're using the tree to store something keyed by
the digest hash of something.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>