Rewrite maths procedures using integer maths rather than floating point
Floating point maths is not accurate enough at the scales required for a
2^256 leaf Sparse Merkle Tree:

-----
Floating point numbers are weird enough to drive anyone bat shit:

251.0
15
251.000000000000028421709430404007434844970703125
-----
  -- From Peter Bex on chicken-users, https://lists.nongnu.org/archive/html/chicken-users/2016-04/msg00010.html

We're only using the integer parts of the results of these procedures so
here we reimplement them in a more robust way and refactor the call sites.

With thanks to Peter Bex and John Cowan on chicken-users in the thread
linked above.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>