Change representation of digest-primitive ahead of tree update mechanics
When we implement merkle-tree-{append,update} as functional operations
on the Merkle Tree, we will need a way to create a new merkle-tree
using the parameters of the given merkle-tree. It is not currently
possible to extract the "digest-primitive" parameter in a way that is
suitable for passing to make-merkle-tree.

This patch moves the evaluation sites around a bit so that the result of
merkle-tree-digest-primitive is symmetric with the corresponding
argument to make-merkle-tree.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>