A Sparse Merkle Tree backend to allow efficient, non-full sparse merkle-trees
The "Sparse Merkle Tree" code has, so far, been tested with the default
dyn-vector backing store implementation. The dyn-vector backing store
allocates space for every possible leaf node which is inefficient for
sparsely populated Merkle Trees.

This patch introduces two new APIs: make-{dense,sparse}-backing-store.

A "Dense" backing store allocates space for every single possible entry
up to the largest leaf number that has been specified so far (i.e. the
size of the Merkle Tree). This kind of store grows to the appropriate
size  when a new "highest leaf number" is seen. The maximum size is
limited by available memory.  We use the existing dyn-vector backing
store as the implementation for the "Dense" backing store.

A "Sparse" backing store only allocates space for the non-default leaves
in the Merkle Tree. The size of a "Sparse" backing store is fixed at
creation time and starts off with all leaves containing the "default"
value. As such, "Sparse" Merkle Trees are always "full". They are
intended to be used in situations that require a "full" tree of
uncomputable size. A "Sparse" Merkle Tree is limited by the number of
non-default leaves in the tree.

This patch introduces code which implements the "Sparse" backing store
in terms of an alist that maps leaf numbers to leaf contents. This has
the advantage of being quick and easy but also has a number of
drawbacks:

 + The alist is unordered so we need to do a linear search of the
   populated leaves in order to find the one we want. i.e it scales
   badly for large numbers of non-default leaves.

 + The backing store implementation contains adhoc support for large
   Leaf IDs. The Leaf ID is a number, counting from 0 from the left hand
   side of the tree. Sparse Merkle Trees, by design, can be uncomputably
   large. Often we want to use the numerical value of, for example, the
   fixed sized output of a digest algorithm such as SHA-256 to index
   into a Merkle Tree. Such a digest would often be represented as
   256bit blob rather than a number. To facilite this, this backing
   store interprets blobs as numbers in Big Endian format.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>