The SQLite backing store written at GDS
This is the work done on the SQL backing store whilst at GDS. It is not
yet complete or finished but it works for simple cases.

The make-sqlite-backing-store procedure has some test code in next-version
that seems to be aimed at exercising the versioning logic in order to test
it.

Signed-off-by: Andy Bennett <andyjpb@digital.cabinet-office.gov.uk>
Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>